asset: cartoon FX 다운
다운 했으요

+++ Assets/2.script/Enemy/Enemy.cs
@@ -62,7 +62,7 @@ public abstract class Enemy : MonoBehaviour
             {
                 float damage = _movespeed * _damage;
 
-                player.TakeDamage(damage);
+                player.TakeDamage((int)damage);
             }
 
             //나죽고
